first attempt at faster GP
we still loop through each observation in the model block. But, the
code should be at least a bit more efficient.

Next step, we will set things up so that we only need to loop through
subjects in the model block, and otherwise give a vector to the
normal() sampling statement.