fake_data2 complete
same as fake_data one except contrast matrix was 0.5, at least for the
subjects (I actually still used a 01 contrast for population effects by
accident). The other difference was that I normalized the trials to a
zero start. Finally, there was much less volatility in the data

The model captures the true group functions accurately!